starting tower placment logic. this involved a bunch of rewriting the pathfinding stuff to make it work with new requirements. Also rearranged many scripts into diferent folders.
