+++ Altair/Assets/Prefabs/Enemes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64cb7b67eece5f24e82771b1079fa7a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
