+++ Altair/Assets/Prefabs/Map.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d81d91d8bdd53324aac0f65fc499af55
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
