+++ Altair/Assets/Prefabs/Map/TilePrefab.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 30beb81b136fbee4b96b6917acc94b88
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
