+++ Altair/Assets/Prefabs/Towers/Tower.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a80401d902bb24e4a965b6cac0049681
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
