+++ Altair/Assets/Scenes/TowerPlacmentTestScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5e7599435e420514e99ba00dc1d70782
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
