+++ Altair/Assets/Scripts/Interface.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a87c06b384bd124092092fa27eeecbd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
