+++ Altair/Assets/Scripts/Interface/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8909f3ddf66731643b97707b21fd7d39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
