+++ Altair/Assets/Scripts/Interface/UI/HUD.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4903ec4564bc764c8c22a487f210f48
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
