+++ Altair/Assets/Scripts/Interface/UI/Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8257ba455daa362498efff718d67b005
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
