+++ Altair/Assets/Scripts/Map/Construction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ef741e62c2c0df40a0c0eb5160640e6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
