+++ Altair/Assets/Scripts/Map/Construction/PrefabSet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8f7b0de69b79dff4b849a81294578bb9
+guid: 76beda47c41828343b48e15adf616649
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
