+++ Altair/Assets/Scripts/Map/Construction/TowerBuildSite.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13bc8dbf0b13a25468615355eb086b88
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
