+++ Altair/Assets/Scripts/Map/PathColor.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using Assets.Scripts.Map.Pathfinding;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
