--- Altair/Assets/Scripts/Map/PathType.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Assets.Scripts.Map
-{
-    public enum PathType
-    {
-        A, B, C
-    }
-}
