+++ Altair/Assets/Scripts/Map/Pathfinding.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c24ad968c3af91c47b735d99a965a1ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
