+++ Altair/Assets/Scripts/Map/Pathfinding/EnemyPathRegen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9aad0b58383e9bc4a89fdb4c5748c3bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
