+++ Altair/Assets/Scripts/Map/Pathfinding/PathNode.cs.meta
