+++ Altair/Assets/Scripts/Map/Pathfinding/PathType.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Assets.Scripts.Map.Pathfinding
+{
+    public enum PathType
+    {
+        A, B, C
+    }
+}
