+++ Altair/Assets/Scripts/Map/Pathfinding/PathType.cs.meta
